requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
SQLAlchemy==2.0.30
sqlmodel==0.0.19
alembic==1.13.1
psycopg2-binary==2.9.9
passlib==1.7.4
bcrypt==4.1.3
oauthlib==3.2.2
python-jose==3.3.0
python-multipart==0.0.9
Jinja2==3.1.4
